refactor: resolve PR comments
